+++ src/scene.h
@@ -11,7 +11,7 @@ struct directional_light {
 
 struct scene {
 	transform_3d view;
-	float perspective;
+	double perspective;
 	rgb_color ambient_light;
 	struct directional_light *directional_lights;
 	int directional_light_count;
